Sem IV/Systems for Design & Implementation/Laboratories/Lab8/backend/bankmanagement/message/serializers.py
from rest_framework import serializers
from .models import Message


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = ["id", "user", "content", "timestamp"]
