Sem IV/Systems for Design & Implementation/Laboratories/Lab8/backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2